+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file starting wit
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2025-x-xx
+
+### Added
+
+-
+
+### Changed
+
+-
+
+### Fixed
+
+- Movie clean title table would not properly save/update user settings or defaults after being modified once.
+
+### Removed
+
+-
+
 ## [0.6.4] - 2025-3-23
 
 ### Added