+++ CHANGELOG.md
@@ -10,12 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Built a new widget that can scroll for very long error messages to replace the default unhandled error box.
+- Logger will now log output to console if debug executable is executed.
 
 ### Changed
 
 - Improve logging in the plugin loader.
 - Use new error window widget to display unhandled errors.
 - Now uses UV instead of Poetry.
+- Logger starts in debug mode and is configured on first message via the users settings.
 
 ### Fixed